Renamed and completed Application to DiagramApplication. Write Program class to demonstrate all the AdapterEngine.

+++ structural/Adapter/AdapterEngine/geometrical/rectangular/Square.cs
@@ -6,7 +6,7 @@ internal class Square : Rectangle
 {
     
 
-    public Square(float width) : base(width, width)
+    public Square(double width) : base(width, width)
     {
         
     }
